requirements.txt
flask==2.0.3
werkzeug==2.0.3
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.2
tensorflow==2.6.2
xgboost==1.5.2
joblib==1.1.1
matplotlib==3.3.4
seaborn==0.11.2
pymongo==4.1.1
gunicorn==21.2.0
flask-wtf==1.0.1
python-dotenv==0.20.0
bcrypt==4.0.1
PyJWT==2.4.0
Flask-PyMongo==2.3.0
flask-session==0.4.1
email-validator==1.3.1 
